Apply/unapply const here and there

+++ Marlin/src/gcode/config/M200-M205.cpp
@@ -61,7 +61,7 @@ void GcodeSuite::M201() {
 
   LOOP_XYZE(i) {
     if (parser.seen(axis_codes[i])) {
-      const uint8_t a = (i == E_AXIS ? E_AXIS_N(target_extruder) : i);
+      const uint8_t a = (i == uint8_t(E_AXIS) ? E_AXIS_N(target_extruder) : i);
       planner.settings.max_acceleration_mm_per_s2[a] = parser.value_axis_units((AxisEnum)a);
     }
   }
